requirements.txt
Flask>=2.2.0,<3.0.0
requests>=2.28.0
# If/when you use Gemini API, add the official client here
python-dotenv>=1.0.0
google-generativeai>=0.3.0
markdown>=3.4.0
sympy>=1.12
pint>=0.23
flask_sqlalchemy>=3.1.1
